PATİ Mobile Application Project
In this project, the main purpose of which is to create social awareness, a game with a score is also aimed. The project was prepared for TUBITAK and was accepted by TUBITAK. A mobile application was developed using Firebase and Kotlin, which users can easily use and will be rewarded with a prize at the end.